update tests and cleanup

+++ pycromanager/test/conftest.py
@@ -63,7 +63,8 @@ def download_mm_nightly():
     yield mm_installer
 
     # cleanup
-    os.remove(mm_installer)
+    if os.path.isfile(mm_installer):
+        os.remove(mm_installer)
 
 
 @pytest.fixture(scope="session")
@@ -142,14 +143,15 @@ def launch_mm_headless(install_mm):
     config_file = os.path.join(mm_install_dir, 'MMConfig_demo.cfg')
 
     print('Launching Micro-manager in headless mode.')
-    java_loc = None
+    
     # MM doesn't ship with Java on Mac so allow it to be defined here
+    java_loc = None
     if "JAVA" in os.environ:
         java_loc = os.environ["JAVA"]
 
     start_headless(mm_install_dir, config_file, java_loc=java_loc)
 
-    yield
-
-    cleanup()
-    pass
+    # yield
+    #
+    # cleanup()
+    # pass
